Minor refactoring: moved mig complexity code to separate file
